requirements.txt
crewai
crewai-tools
python-dotenv
google-genai